requirements.txt
asgiref==3.4.1
class-registry==2.1.2
Django==3.2.5
django-filter==2.4.0
djangorestframework==3.12.4
filters==1.3.2
Markdown==3.3.4
Pillow==8.3.1
psycopg2==2.9.1
Pygments==2.9.0
python-dateutil==2.8.2
pytz==2021.1
regex==2021.7.6
six==1.16.0
sqlparse==0.4.1
